change *all* debug() calls to the (Symbol, closure) variant
This way, all preparation of debug output will only be done when a
debug point is actually active, and thus not slow down the normal
non-debug flow without need. This results, as it turns out, in a
speedup of about factor 5.

Suddenly, kotlin doesn't feel *so* slow any more -- the difference
between lyk and lingo has just shrunk from a factor of around 17
to a factor of 3.5! And I know there is still room for optimisation
in lyk.