some more 10 to 20 % speed gain by further debug optimisation
Keep track globally if there is any debug topic "on" helps keeping
calls to debug() short, because only one boolean has to be checked,
not fetching a value out of a hash table all the time. This is fun!

Also, a 'cons debug, to see where the (apparently redundant) conses
are called.

UPDATE: make that a 100 %, we are at > 8 million evals/s now.