eliminate eval bug by copying the arglist again on arg eval
Replacing the arguments by their respective eval result directly on
the arglist may have been a nice speed improvement, but it
introduced a very hard to locate eval bug on recursion -- apparently
the arglist is shared in a non-obvious way.

Also,
- all debug topic symbol variables now begin with "debug"
- factored out and improved the frame display for eval debugging
- speaking of which, cleaned up and much improved eval debugging
  (although there is still room for more)
- made the (now single) debug function more flexibly usable -- so
  far it does, more or less, single-handedly eval step debugging
  with little extra code in the closure
- moved more global variables (notably debug topic symbols) to
  globalvars
- some kotlinification: more under_scored_names to camelCase;
  LispError.asObject() => toObject()
- collapsed the with*Environment functions
- fix: defvar is a special form, too (or was it defparameter?)
- cleaned up Streams' print and write functions
- started eval stack dump on error in the repl (still nascent)

This is, of course, a far too big commit. But what should I do when
everything happens so quickly?!