fix botched eval stack in the regtests
I guess the the problem wasresidual parts of stack traces from
caught exceptions, of which there are a lot in the regtests because
every test runs in errset.

Fix: don't use a global variable, but carry the stack along the
caught and re-thrown exceptions. This way the stack is gone when
an exception is discarded. Turns out it works -- not only are the
printed stacks no longer botched, but they are gone, which should
have been like that in the first place.