`(return value` function returns from the innermost function
Not, as assumed earlier, using an implicit block, but simply reuse
the try...finally clause recently introduced for calledFunctionName,
and insert a catch ReturnSignal here. This was easier than expected,
and with no perceptible performance impact with existing Lisp code.

In a second step I learned this should be done only for lambdas,
as otherwise unexpected things happen.

The performance impact of (return value) isn't known, though. I
assume it isn't so bad, in particular as it can avoid ugly and
inefficient Lisp code in some scenarios.