pass a supplied_p Map to the builtin functions
An additional argument Map<LSymbol, Boolean> to the builtin functions
contains the mappings symbol => true for all the supplied optional
and keyword arguments. This is easy, but does come with a performance
cost in the area of 4 %. A set would suffice, but is it cheaper?

Before, the new-factor test was around 28.2 Mill. evals/s; after,
around 27 Mill. That is noticeable, but maybe worth the gain. I am
only afraid it will be more expensive for the lambdas.

Turns out a set is actually *slower* with consistently below 27
Mill. evals/s. So, the Map it is.