keep the lexical envronment in macro
...just like in a function. Using that in macro expansion seems to
be a magical cure for a few macro problems, although I am not quite
sure if that is real or imaginary (certainly complex!).

Got this from CLtL 8.1. Macro Definition. While it is true that in
really desperate cases a look into the documentation *can* help,
this just seems too good to be true. But maybe I'll take another
page from GLS's book (pardon the pun) and do macro *calls* when
they happen, meaning expand and eval. And maybe pre-expand Lisp
code only in defun -- that doesn't sound so bad. Maybe this *is*
the answer I have been looking for.