try to find which paren is not closed at EOF
This wasn't very successful so far, but on top of that we have a
new prompt setting in the interactive Reader, based on *console-prompt*,
which may be a string or function or whatever, and a -V option to
show the version. The latter is done via the build-info function,
so if -V works, that means the engine has been started successfully.