now really implement format ~R, i.e. "english"
Also, related stuff -- regtests, a debug topic ("english"), and a
StrBuf.desc() showing the contents of the StrBuf.

That was fun! Sadly it showed that our "integers" have limited
precision. Not a limited range -- they are Doubles after all! --
but they are imprecise at the end of their range. That isn't good.