make an install package (tarball)
This includes proper staging (in src/stage) and building a tarball
from there, as well as the install target now installing from that
tar ball. The tar ball is relative to $(INSTALLBASE), but OTOH
INSTALLBASE is baked into the scripts. I am not totally happy with
that, but it will have to do for now.

The better solution would be the wrapper scripts deriving INSTALLBASE
from their own call path. Maybe tomorrow.